+++ flask/templates/login.html
@@ -57,7 +57,7 @@
     </style>
   </head>
   <body>
-    <form id="login-form" method="POST" action="/login">
+    <form id="login-form" method="POST" action="/auth/login">
       <h1>Welcome to fpm Registry!</h1>
       <p>Please enter your email and password to log in.</p>
       <input type="email" name="email" placeholder="Email" />
