+++ flask/templates/signup.html
@@ -57,7 +57,7 @@
   </style>
 </head>
 <body>
-  <form id="signup-form" method="POST" action="/signup">
+  <form id="signup-form" method="POST" action="/auth/signup">
     <h1>Sign Up to fpm Registry</h1>
     <p>Please enter your name, email, and password to create an account.</p>
     <input type="text" name="name" placeholder="Name">
